Alteração tipo do atributo peso da Classe Produto
Alteração do tipo do atributo peso da Classe Produto, junto com a inclusão de mais uma linha de comentário explicando a classe

+++ PastaFontesApp/Produto.java
@@ -1,23 +1,25 @@
-// classe que guarda dados de produtos e servicos oferecidos no estacionamento
+// classe de objetos que guarda dados de produtos e servicos oferecidos no estacionamento
+// exemplos de produtos e servicos: estacionar veiculo, lavar veiculo, trocar oleo
+
 public class Produto {
 
-	private double peso;
+	private int peso;
 	private double altura;
 
 	public Produto() {
 
 	}
 
-	public Produto(double peso, double altura) {
+	public Produto(int peso, double altura) {
 		this.peso = peso;
 		this.altura = altura;
 	}
 
-	public double getPeso() {
+	public int getPeso() {
 		return peso;
 	}
 
-	public void setPeso(double peso) {
+	public void setPeso(int peso) {
 		this.peso = peso;
 	}
